Refactor to begin separating concerns - determining package list & running audits & writing results now in distinct modules
