+++ audit/docker/README.md
@@ -1,0 +1,3 @@
+To avoid the host machine getting pwned, we run the actual build step inside a Docker container - see auditAll.js for context. All code related to doing that lives in this folder.
+
+Currently, this folder is self-contained; that is, `docker/buildPackage.js` does not import anything from outside the `docker/` folder, and nothing from outside the `docker/` folder gets copied into the Docker image when it is built. This is not the only possible design, and has the obvious *disadvantage* that code cannot be shared between the script that runs in Docker and the script that runs on the host machine; any common utility functions must instead be duplicated. The *advantage*, though, is that when writing code outside `docker/`, we can be sure it will run only on the host machine, and do not have to wonder whether it is Docker-compatible or whether we are inefficiently going to repeat steps inside Docker that were already cached on the host machine. At the time of writing, this seems like a good tradeoff.
